refactor(composition): split parse.ts into parse.ts and format.ts
