+++ src/lib/composition/Formula.svelte
@@ -5,13 +5,16 @@
   import ElementTile from '$lib/element/ElementTile.svelte'
   import { format_num } from '$lib/labels'
   import type { HTMLAttributes } from 'svelte/elements'
-  import type { ElementWithOxidation, OxiComposition } from './parse'
   import {
     format_oxi_state,
-    oxi_composition_to_elements,
-    parse_formula_with_oxidation,
     sort_by_electronegativity,
     sort_by_hill_notation,
+  } from './format'
+  import {
+    type ElementWithOxidation,
+    oxi_composition_to_elements,
+    type OxiComposition,
+    parse_formula_with_oxidation,
   } from './parse'
 
   type FormulaOrdering = `electronegativity` | `alphabetical` | `original` | `hill`
