+++ src/lib/composition/index.ts
@@ -4,6 +4,7 @@ export { default as BarChart } from './BarChart.svelte'
 export { default as BubbleChart } from './BubbleChart.svelte'
 export { default as Composition } from './Composition.svelte'
 export { default as Formula } from './Formula.svelte'
+export * from './format'
 export * from './parse'
 export { default as PieChart } from './PieChart.svelte'
