+++ src/lib/phase-diagram/thermodynamics.ts
@@ -3,7 +3,7 @@ import {
   count_atoms_in_composition,
   extract_formula_elements,
   sort_by_electronegativity,
-} from '$lib/composition/parse'
+} from '$lib/composition'
 import * as math from '$lib/math'
 import {
   barycentric_to_ternary_xyz,
